+++ src/Controller/CommandeProductController.php
@@ -94,35 +94,4 @@ class CommandeProductController extends AbstractController
 
         return $this->redirectToRoute('commande_product_index');
     }
-
-    /**
-     * @Route("/panier", name="cart_index")
-     */
-    public function panier(CartService $cartService): Response
-    {
-        return $this->render('components/pages/cart/index.html.twig', [
-            'items' => $cartService->getFullCart(),
-            'total' => $cartService->getTotal()
-        ]);
-    }
-
-    /**
-     * @Route("/panier/add/{id}", name="cart_add")
-     */
-    public function add($id, CartService $cartService) {
-        
-        $cartService->add($id);
-
-        return $this->redirectToRoute("cart_index");
-    }
-
-    /**
-     * @Route("/panier/remove/{id}", name="cart_remove")
-     */
-    public function remove($id, CartService $cartService) {
-
-        $cartService->remove($id);
-
-        return $this->redirectToRoute("cart_index");
-    }
 }
