+++ src/Controller/DefaultController.php
@@ -9,9 +9,6 @@ use App\Repository\ProducerRepository;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-
-
-
 class DefaultController extends AbstractController
 {
     /**
