+++ src/Controller/PanierController.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace App\Controller;
+
+use App\Service\Cart\CartService;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+
+class PanierController extends AbstractController
+{
+    /**
+     * @Route("/panier", name="cart_index")
+     */
+    public function panier(CartService $cartService): Response
+    {
+        return $this->render('components/pages/panier/index.html.twig', [
+            'items' => $cartService->getFullCart(),
+            'total' => $cartService->getTotal()
+        ]);
+    }
+
+    /**
+     * @Route("/panier/add/{id}", name="cart_add")
+     */
+    public function add($id, CartService $cartService) {
+        
+        $cartService->add($id);
+
+        return $this->redirectToRoute("cart_index");
+    }
+
+    /**
+     * @Route("/panier/remove/{id}", name="cart_remove")
+     */
+    public function remove($id, CartService $cartService) {
+
+        $cartService->remove($id);
+
+        return $this->redirectToRoute("cart_index");
+    }
+}
