+++ src/Controller/UserController.php
@@ -59,7 +59,7 @@ class UserController extends AbstractController
 
                     // Move file
                     $uploaded_file->move(
-                        "./upload/",
+                        "./profile/",
                         $new_file
                     );
